chore: combine all into main script
- for the time being, keep everything maintained into main.py
- then, once the project is a bit more stable, separate into modules
